PDFBOX-4071: remove uneeded unboxing
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1823365 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/PDPageContentStream.java
@@ -449,7 +449,7 @@ public final class PDPageContentStream implements Closeable
             }
             else if (obj instanceof Float)
             {
-                writeOperand(((Float) obj).floatValue());
+                writeOperand((Float) obj);
             }
             else
             {
